% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{diagnostic_pca_cloud}
\alias{diagnostic_pca_cloud}
\title{Plot PCA Diagnostic Scatter Plot}
\usage{
diagnostic_pca_cloud(
  original_features,
  new_features,
  PC_a = 1,
  PC_b = 2,
  save = NULL
)
}
\arguments{
\item{original_features}{A matrix of PCA-transformed original features.}

\item{new_features}{A matrix of PCA-transformed new features.}

\item{PC_a}{Integer. The first principal component to plot on the x-axis. Default is 1.}

\item{PC_b}{Integer. The second principal component to plot on the y-axis. Default is 2.}

\item{save}{Character. Optional file path to save the plot. Default is NULL.}
}
\value{
A ggplot2 object showing the scatter plot.
}
\description{
This function generates a scatter plot comparing the original PCA-transformed features
with a new set of PCA-transformed features. The original features are plotted in grey,
while the new features are plotted in red.
}
\examples{
\dontrun{
diagnostic_pca_cloud(original_features, new_features, PC_a = 1, PC_b = 2, save = "pca_plot.png")
}

}
