% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_cda_model.R
\name{extract_features}
\alias{extract_features}
\title{Extract features from the last pooling layer of a Keras model}
\usage{
extract_features(model = "base_cnn", images)
}
\arguments{
\item{model}{A Keras model from which to extract features.}

\item{images}{A 4D array of images (height, width, channels, num_images).}
}
\value{
A matrix containing the extracted features.
}
\description{
This function extracts features from the last pooling layer of the model
for a set of images.
}
