% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_cdascorer_dataframe.R
\name{load_cdascorer_dataframe}
\alias{load_cdascorer_dataframe}
\title{Load a CSV file for AutoCDAScorer Dataframe}
\usage{
load_cdascorer_dataframe(filepath)
}
\arguments{
\item{filepath}{A string representing the path to the CSV file.}
}
\value{
A dataframe containing the data from the CSV file if it exists and contains the required columns.
}
\description{
This function loads a CSV file into a dataframe, ensuring that the file exists and contains the required columns: "img", "x1", "x2", "y1", and "y2".
}
\examples{
df <- load_cdascorer_dataframe("path/to/data.csv")
}
