% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_functions.R
\name{load_images_and_labels}
\alias{load_images_and_labels}
\title{Load images and file names from a directory, with optional labels}
\usage{
load_images_and_labels(data_dir, image_size = 64, labels = FALSE)
}
\arguments{
\item{data_dir}{A string specifying the directory containing images.}

\item{image_size}{An integer specifying the size to which each image should be resized (width, height). Default is 64.}

\item{labels}{A logical indicating whether to assign labels based on subdirectory names. Default is FALSE.}
}
\value{
A list containing:
\item{images}{A 4D array of images: n_images, height, width, channels}
\item{labels}{(Only if labels = TRUE) A vector of class labels corresponding to each image}
\item{filenames}{A vector of image file names}
}
\description{
Loads all \code{.tif} images from the given directory, with the option to assign labels.
If \code{labels = TRUE}, images are expected to be inside first-level subdirectories, and the subdirectory names are used as labels.
If \code{labels = FALSE}, images are collected from both the main directory and first-level subdirectories, and no labels are assigned.
}
