% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{pca_diagnostic_convexhull}
\alias{pca_diagnostic_convexhull}
\title{Scatter Plot of PCA-Transformed Features with Convex Hull}
\usage{
pca_diagnostic_convexhull(pca, new_features, PC_a, PC_b)
}
\arguments{
\item{pca}{A PCA object loaded from package data.}

\item{new_features}{A matrix of new features transformed by PCA.}

\item{PC_a}{An integer specifying the first principal component to plot.}

\item{PC_b}{An integer specifying the second principal component to plot.}
}
\value{
A ggplot2 object representing the PCA scatter plot with a convex hull.
}
\description{
This function creates a scatter plot of PCA-transformed features, highlighting the convex hull of
the original data points. It also classifies new data points as either inside or outside the convex hull.
}
