% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{pca_diagnostic_density}
\alias{pca_diagnostic_density}
\title{Scatter Plot of PCA-Transformed Features with Density Contours}
\usage{
pca_diagnostic_density(pca, new_features, PC_a, PC_b, num_bins = 3)
}
\arguments{
\item{pca}{A PCA object loaded from package data.}

\item{new_features}{A matrix of new features transformed by PCA.}

\item{PC_a}{An integer indicating the index of the first principal component for plotting.}

\item{PC_b}{An integer indicating the index of the second principal component for plotting.}

\item{num_bins}{An integer specifying the number of contour bins for the density plot (default: 3).}
}
\value{
A ggplot2 object representing the PCA scatter plot with density contours.
}
\description{
This function creates a scatter plot of original and new features transformed by PCA,
overlaying density contours to visualize original data concentration using kernel density estimation (KDE).
}
