% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{pca_diagnostic_target}
\alias{pca_diagnostic_target}
\title{Scatter Plot of PCA-Transformed Features with Confidence Ellipses}
\usage{
pca_diagnostic_target(pca, new_features, PC_a, PC_b, num_ellipses = 3)
}
\arguments{
\item{pca}{A PCA object loaded from package data.}

\item{new_features}{A matrix of new features transformed by the same PCA.}

\item{PC_a}{An integer indicating the index of the first principal component for plotting.}

\item{PC_b}{An integer indicating the index of the second principal component for plotting.}

\item{num_ellipses}{An integer specifying the number of ellipses to draw for visualizing data spread (default: 3).}
}
\value{
A ggplot2 object representing the PCA scatter plot with confidence ellipses.
}
\description{
This function creates a scatter plot of original and new features transformed by PCA,
highlighting the distribution of data points along two selected principal components.
It also overlays confidence ellipses to visualize the data spread.
}
