% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{pca_plot_with_convex_hull}
\alias{pca_plot_with_convex_hull}
\title{Scatter Plot of PCA-Transformed Features with Convex Hull}
\usage{
pca_plot_with_convex_hull(original_features, new_features, PC_a = 1, PC_b = 2)
}
\arguments{
\item{original_features}{A matrix of original features transformed by PCA.}

\item{new_features}{A matrix of new features transformed by PCA.}

\item{PC_a}{An integer specifying the first principal component to plot (default: 1).}

\item{PC_b}{An integer specifying the second principal component to plot (default: 2).}
}
\value{
A ggplot2 object representing the PCA scatter plot with a convex hull.
}
\description{
This function creates a scatter plot of PCA-transformed features, highlighting the convex hull of
the original data points. It also classifies new data points as either inside or outside the convex hull.
}
