% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{pca_transform_features}
\alias{pca_transform_features}
\title{Apply a pre-trained PCA transformation to new features}
\usage{
pca_transform_features(pca, features)
}
\arguments{
\item{pca}{A prcomp object containing the trained PCA model.}

\item{features}{A matrix or array of new features to transform.}
}
\value{
A matrix of PCA-transformed features.
}
\description{
This function applies a previously trained PCA model to a new set of features without modifying the PCA itself.
}
\examples{
pca_model <- prcomp(existing_images, center = TRUE, scale. = TRUE)
transformed_images <- pca_transform_images(pca_model, new_images)
}
