% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{pca_transform_python}
\alias{pca_transform_python}
\title{Apply pre-trained PCA transformation to new data using Python}
\usage{
pca_transform_python(pca, data)
}
\arguments{
\item{pca}{A list returned by \code{run_pca_python}, containing a trained PCA model.}

\item{data}{An array of extracted features or a list containing an array of images.}
}
\value{
A matrix of PCA-transformed features.
}
\description{
This function applies a previously trained PCA model to a new dataset without modifying the PCA itself.
}
