% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_functions.R
\name{rgb_to_bgr}
\alias{rgb_to_bgr}
\title{Convert RGB to BGR}
\usage{
rgb_to_bgr(data)
}
\arguments{
\item{data}{A dataset containing images, where images are stored in a 4D array, where the 4th dimension is RGB colour}
}
\value{
A dataset whose images have the same structure, but the color channels reordered to BGR.
}
\description{
This function takes a converts a 4D image array from RGB to BGR by reordering the color channels.
}
