% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{run_pca}
\alias{run_pca}
\title{Perform PCA}
\usage{
run_pca(data, n_components = 10, savepath = NULL)
}
\arguments{
\item{data}{An array of extracted features or a list containing an array of images.}

\item{n_components}{Integer. The number of principal components to retain. Default is 10.}

\item{savepath}{Character. Optional file path to save the PCA model as an RDS file. Default is NULL.}
}
\value{
A list containing:
\item{pca}{A \code{prcomp} object containing the trained PCA model.}
\item{features_pca}{A matrix of transformed features after applying PCA.}
\item{explained_variance}{Proportion of variance explained by each principal component.}
\item{principal_components}{The principal component vectors.}
}
\description{
This function applies Principal Component Analysis (PCA) to pixel values or extracted features.
Optionally, it can save the PCA model as an RDS file.
}
