% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{run_pca_python}
\alias{run_pca_python}
\title{Perform PCA in Python}
\usage{
run_pca_python(data, n_components = 10)
}
\arguments{
\item{data}{An array of extracted features or a list containing an array of images.}

\item{n_components}{Integer. The number of principal components to retain. Default is 10.}
}
\value{
A list containing:
\item{pca}{The trained PCA model from \code{sklearn.decomposition.PCA}.}
\item{features_pca}{The dataset transformed into the principal component space.}
\item{explained_variance}{Proportion of variance explained by each principal component.}
\item{principal_components}{The principal component vectors.}
}
\description{
This function applies Principal Component Analysis (PCA) to pixel values or extracted features.
}
