% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{transform_features_python}
\alias{transform_features_python}
\title{Transform New Data Using a Trained PCA Model}
\usage{
transform_features_python(dataset, pca)
}
\arguments{
\item{dataset}{A list containing an \code{images} element, which is a 3D array (e.g., \link{samples, height, width}).}

\item{pca}{A list returned by \code{run_pca_python}, containing a trained PCA model.}
}
\value{
A matrix of transformed features in the principal component space.
}
\description{
This function applies a previously trained PCA model to transform new dataset features.
}
