% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_functions.R
\name{annotations_to_csv}
\alias{annotations_to_csv}
\title{Save annotations to CSV}
\usage{
annotations_to_csv(data, predictions, output_path)
}
\arguments{
\item{data}{A list containing a 4D array of images (batch, height, width, channels) and a vector of filenames.}

\item{predictions}{A vector of predicted class labels or a matrix of raw model predictions.}

\item{output_path}{The file path where the CSV file should be saved.}
}
\value{
Writes a CSV file with two columns: "name" and "prediction".
}
\description{
This function saves the predicted annotations (image filepaths and predicted classes) to a CSV file.
}
