% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keras_functions.R
\name{load_cda_model}
\alias{load_cda_model}
\title{Load CDAScorer Keras Model}
\usage{
load_cda_model(model)
}
\arguments{
\item{model}{A string specifying which model to load.}
}
\value{
A Keras model object
}
\description{
This function loads a pre-trained Keras model from the \code{extdata} directory of the \code{AutoCDAScorer} package.
}
