% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_cda_model.R
\name{load_cda_model}
\alias{load_cda_model}
\title{Load Keras model}
\usage{
load_cda_model()
}
\value{
A Keras model object
}
\description{
This function loads a pre-trained Keras model from the \code{extdata} directory of the \code{AutoCDAScorer} package.
}
\examples{
model <- load_cda_model()

}
