% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_functions.R
\name{load_cdascorer_dataframe}
\alias{load_cdascorer_dataframe}
\title{Load CSV file in CDAScorer format}
\usage{
load_cdascorer_dataframe(filepath)
}
\arguments{
\item{filepath}{A string representing the path to the CSV file.}
}
\value{
A dataframe containing the data from the CSV file.
}
\description{
This function loads a CSV file into a dataframe, ensuring that the file exists and contains the required columns: "img", "x1", "x2", "y1", and "y2".
}
