% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_functions.R
\name{load_images}
\alias{load_images}
\title{Load images and file names from a directory}
\usage{
load_images(input_path, image_size = 64)
}
\arguments{
\item{input_path}{A string specifying the directory containing images.}

\item{image_size}{An integer specifying the size to which each image should be resized (width, height). Default is 64.}
}
\value{
A list containing:
\item{images}{A 4D array of images: n_images, height, width, channels}
\item{filenames}{A vector of image file names}
}
\description{
Loads all \code{.tif} images from the given directory and immediate subdirectories.
}
