% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_cdascorer_dataframe.R
\name{load_images_and_labels}
\alias{load_images_and_labels}
\title{Load images, labels, and file names from a directory}
\usage{
load_images_and_labels(data_dir, image_size = 64)
}
\arguments{
\item{data_dir}{A string specifying the directory containing subdirectories of images.}

\item{image_size}{An integer specifying the size to which each image should be resized (width, height).
Default is 64.}
}
\value{
A list containing:
\item{images}{A 4D array of images}
\item{labels}{A vector of class labels corresponding to each image}
\item{names}{A vector of image file names}
}
\description{
Loads images, labels, and file names from a given directory. Each class of images is assumed
to be stored in subdirectories inside \code{data_dir}, and each subdirectory is considered a separate class.
}
\examples{
\dontrun{
images <- load_images_and_labels(data_dir, image_size=128)
}

}
