% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{load_result_pca}
\alias{load_result_pca}
\title{Load a Saved PCA Result Object}
\usage{
load_result_pca(data)
}
\arguments{
\item{data}{A character string specifying the object to load. Defaults to \code{"base_cnn_pca"}.}
}
\value{
The loaded R object.
}
\description{
This function loads a saved RDS object.
}
