% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{pca_transform}
\alias{pca_transform}
\title{Apply pre-trained PCA transformation to new data}
\usage{
pca_transform(data, pca = NULL)
}
\arguments{
\item{data}{An array of extracted features or a list containing an array of images.}

\item{pca}{A list containing a prcomp pca object.}
}
\value{
A matrix of PCA-transformed features.
}
\description{
This function applies a previously trained PCA model to a new dataset without modifying the PCA itself.
}
