% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_cda_model.R
\name{predict_score}
\alias{predict_score}
\title{Predict the score or class for a batch of images using a Keras model}
\usage{
predict_score(model = "base_cnn", images, softmax = TRUE)
}
\arguments{
\item{model}{A Keras model used for making predictions.}

\item{images}{A 4D array of images (height, width, channels, num_images).}

\item{softmax}{A logical value indicating whether to return raw softmax values
(TRUE) or the predicted class index (FALSE). Default is TRUE.}
}
\value{
A vector or matrix containing either the raw softmax values or the predicted class indices.
}
\description{
This function returns either the raw softmax probabilities or the predicted class
for a set of images.
}
