% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_cda_model.R
\name{rgb_to_bgr}
\alias{rgb_to_bgr}
\title{Convert RGB to BGR}
\usage{
rgb_to_bgr(dataset)
}
\arguments{
\item{dataset}{A dataset containing images, where images are stored in an array with dimensions \verb{[n, m, 3]}, where the 3rd dimension represents the RGB color channels.}
}
\value{
A dataset with the same structure, but the color channels reordered to BGR.
}
\description{
This function takes a dataset containing images in RGB format and converts them to BGR format by reordering the color channels.
}
