% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_cdascorer_dataframe.R
\name{show_test_image}
\alias{show_test_image}
\title{Display a Test Image from the Dataset}
\usage{
show_test_image(dataset)
}
\arguments{
\item{dataset}{A list containing an \code{images} array. The first image in the array will be displayed.}
}
\value{
A visual output of the first image in the dataset.
}
\description{
This function displays the first image from the given dataset using \code{rasterGrob} and \code{grid.draw} from the \code{grid} package.
}
